refactor: 메시지 생성 로직 분리
