+++ fe/src/Components/MessageScreen.jsx
@@ -1,35 +1,24 @@
 import { INIT_ALERT_MESSAGE } from "Helper/constant";
-import { getWonTemplate } from "Helper/utils";
+import addMessageList from "Helper/message";
 import useAlertMessage from "Hooks/useAlertMessage";
-import { useCallback, useEffect, useState } from "react";
+import useMessageList from "Hooks/useMessageList";
+import { useEffect, useRef } from "react";
 import { Message, Screen, ScreenContainer } from "./MessageScreen.styled";
 
 export default function MessageScreen() {
-  const [alertMessage, setAlertMessage] = useAlertMessage();
-  const [messageList, setMessageList] = useState([]);
-
-  const createNewMessageList = useCallback(
-    (message) => {
-      const newMessageList = [...messageList, message];
-      return newMessageList;
-    },
-    [messageList]
-  );
+  const [alertMessage, setAlertMessage] = useAlertMessage({});
+  const [messageList, setMessageList] = useMessageList([]);
+  const scrollRef = useRef();
 
   useEffect(() => {
-    const message = Object.entries(alertMessage).reduce(getAlertMessage, "");
-    if (!message) {
-      return;
-    }
-
-    const newMessageList = createNewMessageList(message);
-    setMessageList(newMessageList);
+    addMessageList({ alertMessage, setMessageList, messageList });
     setAlertMessage(INIT_ALERT_MESSAGE);
-  }, [alertMessage, setAlertMessage, createNewMessageList]);
+    scrollRef.current.scrollIntoView({ behavior: "smooth", block: "end", inline: "nearest" });
+  }, [alertMessage, messageList, setMessageList, setAlertMessage]);
 
   return (
     <ScreenContainer>
-      <Screen>
+      <Screen ref={scrollRef}>
         {messageList &&
           messageList.map((message, idx) => (
             <Message key={createArrayKeyForNoHasID(message, idx)}>{message}</Message>
@@ -38,39 +27,6 @@ export default function MessageScreen() {
     </ScreenContainer>
   );
 }
-
-const getAlertMessage = (message, [alertName, alertAmount]) => {
-  if (!alertAmount) {
-    return message;
-  }
-  return generateMessage(alertName, alertAmount);
-};
-
-const generateMessage = (alertName, alertAmount) => {
-  const alerts = {
-    chargeCash: generateInvestmentMessage,
-    changeAmount: generateChangeMessage,
-    orderTitle: generateOrderMessage,
-    wrong: generateWrongMessage,
-  };
-  return alerts[alertName](alertAmount);
-};
-
-const generateOrderMessage = (title) => {
-  return `${title}가 주문되었습니다.`;
-};
-
-const generateChangeMessage = (amount) => {
-  return `거스름돈 ${getWonTemplate(amount)}이 반환되었습니다.`;
-};
-
-const generateInvestmentMessage = (amount) => {
-  return `자판기에 ${getWonTemplate(amount)}이 투입되었습니다.`;
-};
-
-const generateWrongMessage = () => {
-  return `주문할 수 없는 상품입니다.`;
-};
 
 const createArrayKeyForNoHasID = (message, key) => {
   return `${message + key}`;
